+++ pkg/ingester/ingester.go
@@ -85,6 +85,7 @@ const (
 	instanceIngestionRateTickInterval = time.Second
 
 	sampleOutOfOrder     = "sample-out-of-order"
+	sampleTooOld         = "sample-too-old"
 	newValueForTimestamp = "new-value-for-timestamp"
 	sampleOutOfBounds    = "sample-out-of-bounds"
 )
@@ -122,7 +123,7 @@ type Config struct {
 	ActiveSeriesMetricsIdleTimeout  time.Duration                     `yaml:"active_series_metrics_idle_timeout" category:"advanced"`
 	ActiveSeriesCustomTrackers      activeseries.CustomTrackersConfig `yaml:"active_series_custom_trackers" doc:"description=[Deprecated] This config has been moved to the limits config, please set it there. Additional custom trackers for active metrics. If there are active series matching a provided matcher (map value), the count will be exposed in the custom trackers metric labeled using the tracker name (map key). Zero valued counts are not exposed (and removed when they go back to zero)." category:"advanced"`
 
-	ExemplarsUpdatePeriod time.Duration `yaml:"exemplars_update_period" category:"experimental"`
+	TSDBConfigUpdatePeriod time.Duration `yaml:"tsdb_config_update_period" category:"experimental"`
 
 	BlocksStorageConfig         mimir_tsdb.BlocksStorageConfig `yaml:"-"`
 	StreamChunksWhenUsingBlocks bool                           `yaml:"-" category:"advanced"`
@@ -150,7 +151,7 @@ func (cfg *Config) RegisterFlags(f *flag.FlagSet, logger log.Logger) {
 	f.DurationVar(&cfg.ActiveSeriesMetricsIdleTimeout, "ingester.active-series-metrics-idle-timeout", 10*time.Minute, "After what time a series is considered to be inactive.")
 
 	f.BoolVar(&cfg.StreamChunksWhenUsingBlocks, "ingester.stream-chunks-when-using-blocks", true, "Stream chunks from ingesters to queriers.")
-	f.DurationVar(&cfg.ExemplarsUpdatePeriod, "ingester.exemplars-update-period", 15*time.Second, "Period with which to update per-tenant max exemplar limit.")
+	f.DurationVar(&cfg.TSDBConfigUpdatePeriod, "ingester.tsdb-config-update-period", 15*time.Second, "Period with which to update the per-tenant TSDB configuration.")
 
 	cfg.DefaultLimits.RegisterFlags(f)
 
@@ -413,8 +414,8 @@ func (i *Ingester) updateLoop(ctx context.Context) error {
 	ingestionRateTicker := time.NewTicker(instanceIngestionRateTickInterval)
 	defer ingestionRateTicker.Stop()
 
-	exemplarUpdateTicker := time.NewTicker(i.cfg.ExemplarsUpdatePeriod)
-	defer exemplarUpdateTicker.Stop()
+	tsdbUpdateTicker := time.NewTicker(i.cfg.TSDBConfigUpdatePeriod)
+	defer tsdbUpdateTicker.Stop()
 
 	var activeSeriesTickerChan <-chan time.Time
 	if i.cfg.ActiveSeriesMetricsEnabled {
@@ -441,8 +442,8 @@ func (i *Ingester) updateLoop(ctx context.Context) error {
 			}
 			i.tsdbsMtx.RUnlock()
 
-		case <-exemplarUpdateTicker.C:
-			i.applyExemplarsSettings()
+		case <-tsdbUpdateTicker.C:
+			i.applyTSDBSettings()
 
 		case <-activeSeriesTickerChan:
 			i.updateActiveSeries(time.Now())
@@ -495,14 +496,21 @@ func (i *Ingester) updateActiveSeries(now time.Time) {
 	}
 }
 
-// Go through all tenants and apply the current max-exemplars setting.
-// If it changed, tsdb will resize the buffer; if it didn't change tsdb will return quickly.
-func (i *Ingester) applyExemplarsSettings() {
+// applyTSDBSettings goes through all tenants and applies
+// * The current max-exemplars setting. If it changed, tsdb will resize the buffer; if it didn't change tsdb will return quickly.
+// * The current out-of-order time window. If it changes from 0 to >0, then a new Write-Behind-Log gets created for that tenant.
+func (i *Ingester) applyTSDBSettings() {
 	for _, userID := range i.getTSDBUsers() {
 		globalValue := i.limits.MaxGlobalExemplarsPerUser(userID)
 		localValue := i.limiter.convertGlobalToLocalLimit(userID, globalValue)
-		// We populate a Config struct with just one value, which is OK
-		// because Head.ApplyConfig only looks at one value.
+
+		oooTW := i.limits.OutOfOrderTimeWindow(userID)
+		if oooTW < 0 {
+			oooTW = 0
+		}
+
+		// We populate a Config struct with just TSDB related config, which is OK
+		// because DB.ApplyConfig only looks at the specified config.
 		// The other fields in Config are things like Rules, Scrape
 		// settings, which don't apply to Head.
 		cfg := promcfg.Config{
@@ -510,13 +518,16 @@ func (i *Ingester) applyExemplarsSettings() {
 				ExemplarsConfig: &promcfg.ExemplarsConfig{
 					MaxExemplars: int64(localValue),
 				},
+				TSDBConfig: &promcfg.TSDBConfig{
+					OutOfOrderAllowance: time.Duration(oooTW).Milliseconds(),
+				},
 			},
 		}
-		tsdb := i.getTSDB(userID)
-		if tsdb == nil {
+		db := i.getTSDB(userID)
+		if db == nil {
 			continue
 		}
-		if err := tsdb.db.ApplyConfig(&cfg); err != nil {
+		if err := db.db.ApplyConfig(&cfg); err != nil {
 			level.Error(i.logger).Log("msg", "failed to apply config to TSDB", "user", userID, "err", err)
 		}
 	}
@@ -599,6 +610,7 @@ func (i *Ingester) PushWithCleanup(ctx context.Context, req *mimirpb.WriteReques
 		startAppend               = time.Now()
 		sampleOutOfBoundsCount    = 0
 		sampleOutOfOrderCount     = 0
+		sampleTooOldCount         = 0
 		newValueForTimestampCount = 0
 		perUserSeriesLimitCount   = 0
 		perMetricSeriesLimitCount = 0
@@ -620,12 +632,17 @@ func (i *Ingester) PushWithCleanup(ctx context.Context, req *mimirpb.WriteReques
 			otlog.Int("numseries", len(req.Timeseries)))
 	}
 
+	oooTW := i.limits.OutOfOrderTimeWindow(userID)
 	for _, ts := range req.Timeseries {
 		// The labels must be sorted (in our case, it's guaranteed a write request
 		// has sorted labels once hit the ingester).
 
-		// Fast path in case we only have samples and they are all out of bounds.
-		if minAppendTimeAvailable && len(ts.Samples) > 0 && len(ts.Exemplars) == 0 && allOutOfBounds(ts.Samples, minAppendTime) {
+		// Fast path in case we only have samples and they are all out of bound
+		// and out-of-order support is not enabled.
+		// TODO(jesus.vazquez) If we had too many old samples we might want to
+		// extend the fast path to fail early.
+		if oooTW <= 0 && minAppendTimeAvailable &&
+			len(ts.Samples) > 0 && len(ts.Exemplars) == 0 && allOutOfBounds(ts.Samples, minAppendTime) {
 			failedSamplesCount += len(ts.Samples)
 			sampleOutOfBoundsCount += len(ts.Samples)
 
@@ -676,6 +693,11 @@ func (i *Ingester) PushWithCleanup(ctx context.Context, req *mimirpb.WriteReques
 			case storage.ErrOutOfOrderSample:
 				sampleOutOfOrderCount++
 				updateFirstPartial(func() error { return newIngestErrSampleOutOfOrder(model.Time(s.TimestampMs), ts.Labels) })
+				continue
+
+			case storage.ErrTooOldSample:
+				sampleTooOldCount++
+				updateFirstPartial(func() error { return newIngestErrSampleTooOld(model.Time(s.TimestampMs), ts.Labels) })
 				continue
 
 			case storage.ErrDuplicateSampleForTimestamp:
@@ -780,6 +802,9 @@ func (i *Ingester) PushWithCleanup(ctx context.Context, req *mimirpb.WriteReques
 	}
 	if sampleOutOfOrderCount > 0 {
 		validation.DiscardedSamples.WithLabelValues(sampleOutOfOrder, userID).Add(float64(sampleOutOfOrderCount))
+	}
+	if sampleTooOldCount > 0 {
+		validation.DiscardedSamples.WithLabelValues(sampleTooOld, userID).Add(float64(sampleTooOldCount))
 	}
 	if newValueForTimestampCount > 0 {
 		validation.DiscardedSamples.WithLabelValues(newValueForTimestamp, userID).Add(float64(newValueForTimestampCount))
@@ -1453,6 +1478,7 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 	}
 
 	maxExemplars := i.limiter.convertGlobalToLocalLimit(userID, i.limits.MaxGlobalExemplarsPerUser(userID))
+	oooTW := time.Duration(i.limits.OutOfOrderTimeWindow(userID))
 	// Create a new user database
 	db, err := tsdb.Open(udir, userLogger, tsdbPromReg, &tsdb.Options{
 		RetentionDuration:              i.cfg.BlocksStorageConfig.TSDB.Retention.Milliseconds(),
@@ -1473,8 +1499,11 @@ func (i *Ingester) createTSDB(userID string) (*userTSDB, error) {
 		IsolationDisabled:              !i.cfg.BlocksStorageConfig.TSDB.IsolationEnabled,
 		HeadChunksWriteQueueSize:       i.cfg.BlocksStorageConfig.TSDB.HeadChunksWriteQueueSize,
 		NewChunkDiskMapper:             i.cfg.BlocksStorageConfig.TSDB.NewChunkDiskMapper,
-		AllowOverlappingQueries:        i.cfg.BlocksStorageConfig.TSDB.AllowOverlappingQueries,
-		AllowOverlappingCompaction:     false, // always false since Mimir only uploads lvl 1 compacted blocks
+		AllowOverlappingQueries:        true,                 // We can have overlapping blocks from past or out-of-order enabled during runtime.
+		AllowOverlappingCompaction:     false,                // always false since Mimir only uploads lvl 1 compacted blocks
+		OutOfOrderAllowance:            oooTW.Milliseconds(), // The unit must be same as our timestamps.
+		OutOfOrderCapMin:               int64(i.cfg.BlocksStorageConfig.TSDB.OutOfOrderCapMin),
+		OutOfOrderCapMax:               int64(i.cfg.BlocksStorageConfig.TSDB.OutOfOrderCapMax),
 	}, nil)
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to open TSDB: %s", udir)
@@ -2081,7 +2110,11 @@ func newIngestErrSampleTimestampTooOld(timestamp model.Time, labels []mimirpb.La
 }
 
 func newIngestErrSampleOutOfOrder(timestamp model.Time, labels []mimirpb.LabelAdapter) error {
-	return newIngestErr(globalerror.SampleOutOfOrder, "the sample has been rejected because another sample with a more recent timestamp has already been ingested and out of order samples are not allowed", timestamp, labels)
+	return newIngestErr(globalerror.SampleOutOfOrder, "the sample has been rejected because another sample with a more recent timestamp has already been ingested and out-of-order samples are not allowed", timestamp, labels)
+}
+
+func newIngestErrSampleTooOld(timestamp model.Time, labels []mimirpb.LabelAdapter) error {
+	return newIngestErr(globalerror.SampleTooOld, "the sample has been rejected because another sample with a more recent timestamp has already been ingested and this sample is beyond the out-of-order time window", timestamp, labels)
 }
 
 func newIngestErrSampleDuplicateTimestamp(timestamp model.Time, labels []mimirpb.LabelAdapter) error {