+++ pkg/util/globalerror/errors.go
@@ -55,6 +55,7 @@ const (
 
 	SampleTimestampTooOld    ID = "sample-timestamp-too-old"
 	SampleOutOfOrder         ID = "sample-out-of-order"
+	SampleTooOld             ID = "sample-too-old"
 	SampleDuplicateTimestamp ID = "sample-duplicate-timestamp"
 	ExemplarSeriesMissing    ID = "exemplar-series-missing"